requirements.txt
flask
flask-sqlalchemy
pymysql
cryptography
werkzeug